+++ client/app/bundles/HelloWorld/components/subscriptions/premium_confirmation_modal.jsx
